basic execution test
